// src/queue/generated/artifacts/operation.ts
/*
 * Copyright (c) Microsoft Corporation. All rights reserved.
 * Licensed under the MIT License. See License.txt in the project root for license information.
 *
 * Code generated by Microsoft (R) AutoRest Code Generator.
 * Changes may cause incorrect behavior and will be lost if the code is regenerated.
 */

export enum Operation {
  Service_SetProperties,
  Service_GetProperties,
  Service_GetStatistics,
  Service_ListQueuesSegment,
  Queue_Create,
  Queue_Delete,
  Queue_GetProperties,
  Queue_GetPropertiesWithHead,
  Queue_SetMetadata,
  Queue_GetAccessPolicy,
  Queue_GetAccessPolicyWithHead,
  Queue_SetAccessPolicy,
  Messages_Dequeue,
  Messages_Clear,
  Messages_Enqueue,
  Messages_Peek,
  MessageId_Update,
  MessageId_Delete,
}
export default Operation;